django-app/requirements.txt
anyascii==0.1.7
asgiref==3.3.4
beautifulsoup4==4.8.2
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
cryptography==41.0.4
defusedxml==0.7.1
Django==3.2.22
django-allauth==0.54.0
django-allauth-2fa==0.8
django-cogwheels==0.3
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-i18nfield==1.9.1
django-modelcluster==5.2
django-otp==1.0.5
django-recaptcha==2.0.6
django-settings-export==1.2.1
django-smtp-ssl==1.0
django-storages[google]==1.12.3
django-taggit==2.0
django-treebeard==4.5.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
google-cloud-secret-manager==2.4.0
google-cloud-storage==2.4.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
importlib-metadata==4.0.1
jdcal==1.4.1
l18n==2020.6.1
mysql-connector-python==8.0.25
mysqlclient==2.0.3
oauthlib==3.1.0
openpyxl==3.0.7
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
paypalrestsdk==1.13.1
phpbridge==0.0.3
Pillow==9.4.0
protobuf==3.18.3
psycopg2-binary==2.9.9
pycparser==2.20
pycurl==7.45.2
PyJWT==2.4.0
pyOpenSSL==23.2.0
pytest==6.2.4
pytest-django==4.4.0
python3-openid==3.2.0
python-json-logger==2.0.7
pytz==2021.1
qrcode==6.1
requests==2.31.0
requests-oauthlib==1.3.0
setdefaultencoding==0.0.0a0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.4
stripe==2.62.0
tablib==3.0.0
telepath==0.1.1
typing-extensions==3.10.0.0
urllib3==1.26.17
wagtail==2.16
wagtail-2fa==1.4.2
wagtail-autocomplete==0.6.2
wagtail-metadata==3.4.0
wagtail-modelchooser==2.3.3
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
zipp==3.4.1
